proxy: move proxyprotocol as a separate package
