+++ pkg/proxy/net/packetio_options.go
@@ -16,6 +16,8 @@ package net
 
 import (
 	"net"
+
+	"github.com/pingcap/TiProxy/pkg/proxy/proxyprotocol"
 )
 
 type PacketIOption = func(*PacketIO)
@@ -31,11 +33,15 @@ func WithWrapError(err error) func(pi *PacketIO) {
 }
 
 // WithRemoteAddr
-var _ net.Addr = &originAddr{}
+var _ proxyprotocol.AddressWrapper = &originAddr{}
 
 type originAddr struct {
 	net.Addr
 	addr string
+}
+
+func (o *originAddr) Unwrap() net.Addr {
+	return o.Addr
 }
 
 func (o *originAddr) String() string {
@@ -46,11 +52,4 @@ func WithRemoteAddr(readdr string, addr net.Addr) func(pi *PacketIO) {
 	return func(pi *PacketIO) {
 		pi.remoteAddr = &originAddr{Addr: addr, addr: readdr}
 	}
-}
-
-func unwrapOriginAddr(addr net.Addr) net.Addr {
-	if oaddr, ok := addr.(*originAddr); ok {
-		return oaddr.Addr
-	}
-	return addr
 }
